fix(OpenStackProvider): Support for deploy using normal user
As NON-admin it was not working. Now it is possible to use existing flavors (that cannot be created) by setting the name in the flavor override. Port security need to be required and set in the VMResource.

BREAKING CHANGE: Port security is disabled by and need to be required by the blueprint